requirements.txt
yt-dlp==2024.9.27
termcolor==2.4.0
pillow==10.4.0
tk==0.1.0